fix: 修复 gRPC 代码导入和类型转换问题
- GameGrpcClient: 添加正确的 proto 生成代码导入
- GameGrpcServer: 添加完整的导入语句
- PlayerService: 修复 seq 字段 Long→int 类型转换
- 编译验证通过

+++ gate-service/src/main/java/com/clawai/gatedemo/gate/service/PlayerService.java
@@ -253,7 +253,7 @@ public class PlayerService {
                 playerId,
                 gameId,
                 message.getMsgType() != null ? message.getMsgType() : "unknown",
-                message.getSeq() != null ? message.getSeq() : 0,
+                message.getSeq() != null ? message.getSeq().intValue() : 0,
                 message.getBody() != null ? message.getBody() : Map.of()
             );
